Change m_rootstate back to reference.
Plus other cleanups.

+++ src/UCTNode.h
@@ -69,7 +69,7 @@ public:
     UCTNode* get_first_child() const;
     UCTNode* get_nopass_child(FastState& state) const;
     const std::vector<node_ptr_t>& get_children() const;
-    int count_nodes() const;
+    size_t count_nodes() const;
     node_ptr_t find_and_take_ownership(const int move);
     node_ptr_t find_and_take_ownership(const GameState& g_new, GameState& g_curr);
